+++ ChangeLog.md
@@ -5,6 +5,7 @@
 ### Added
 - Implement method `myDataset::load()` to load mnist dataset.
 - Add method into the class `myDataset` to get the fields.
+- Add the class `myDatasetIterator` to travel a whole dataset.
 
 ### Changed